* Finished refactoring the library
