+++ backend/src/endpoints/createTag.ts
@@ -2,8 +2,7 @@ import { ITag } from "../models/tag/tag.interface";
 import { Tag } from "../models/tag/tag.model";
 import { IUser } from "../models/user/user.interface";
 import { User } from "../models/user/user.model";
-import { getUserIDFromJWT } from "../lib/getUserIDFromJWT";
-
+import { getUserIDFromJWT } from "../lib/jwt";
 export default async function createTag(req: any, res: any) {
     const userTag: ITag = req.body;
     const user: IUser = await User.findById(getUserIDFromJWT(req.headers.token)).populate('tags_R').exec();
