+++ backend/src/endpoints/deleteDocument.ts
@@ -1,9 +1,9 @@
 import * as fs from 'fs';
 import { IUser } from "../models/user/user.interface";
 import { User } from "../models/user/user.model";
-import { getUserIDFromJWT } from "../lib/getUserIDFromJWT";
 import { Document } from "../models/document/document.model";
-import { generateFilePath } from '../lib/generateFilePath';
+import { getUserIDFromJWT } from '../lib/jwt';
+import { generateFilePath } from '../lib/documentOperations';
 
 export default async function deleteDocument(req, res) {
     const user: IUser = await User.findById(getUserIDFromJWT(req.headers.token)).exec();
