+++ backend/src/endpoints/deleteTag.ts
@@ -1,7 +1,7 @@
 import { IUser } from "../models/user/user.interface";
 import { User } from "../models/user/user.model";
-import { getUserIDFromJWT } from "../lib/getUserIDFromJWT";
 import { Tag } from "../models/tag/tag.model";
+import { getUserIDFromJWT } from "../lib/jwt";
 
 export default async function deleteTag(req, res) {
     const user: IUser = await User.findById(getUserIDFromJWT(req.headers.token)).populate('tags_R').exec();
