+++ backend/src/endpoints/getAllTags.ts
@@ -1,6 +1,6 @@
 import { IUser } from "../models/user/user.interface";
 import { User } from "../models/user/user.model";
-import { getUserIDFromJWT } from "../lib/getUserIDFromJWT";
+import { getUserIDFromJWT } from "../lib/jwt";
 
 export default async function getAllTags(req: any, res: any) {
     const user: IUser = await User.findById(getUserIDFromJWT(req.headers.token)).populate('tags_R').exec();
