--- backend/src/lib/extractFileExtension.ts
@@ -1,3 +1,0 @@
-export function extractFileExtension(fileName: string): string {
-    return fileName.split(".").pop();
-}
