+++ backend/src/lib/userUtils.ts
@@ -1,7 +1,10 @@
-import * as mongoose from "mongoose";
+import * as Mongoose from "mongoose";
 import { User } from "../models/user/user.model";
 
-export async function getNextPrimaryNumber(userId: mongoose.Types.ObjectId) {
+/**
+ * Calculates new primary number and updates it in the database.
+ */
+export async function getNextPrimaryNumber(userId: Mongoose.Types.ObjectId) {
     console.log("UserID=", userId)
     let { nextPrimaryNumber } = await User.findOneAndUpdate({ _id: userId }, { $inc : { nextPrimaryNumber: 1 }}, { new: true });
     return nextPrimaryNumber;
